Seperate Fight! and Main
Main now shows a homescreen, with links to fight and Duel of Wits
respective pages. This is more inline with the goal of what FireStarter is
meant to be. Also added links to Range and Cover and Garbled Transmission
and greyed them out since I have yet to implement them.